+++ RELEASES.md
@@ -12,6 +12,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: introduced PAST (provenance abstract syntax tree), intermediate to Poet and Python (WIP)
 * prov-template: gensym variable, accept null binding, and ensure uniqueness
 * prov-template-compiler: fixed issue with generics in generated code
+* prov-template-compiler: predecessor table initialisation
 
 `2.2.2`
 * prov-template: added ptm-merging template